requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
coloredlogs==15.0.1
decorator==5.1.1
environs==11.2.1
flatbuffers==24.3.25
frozenlist==1.5.0
humanfriendly==10.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.43.0
magic-filter==1.0.12
marshmallow==3.23.1
moviepy==2.0.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
onnxruntime==1.20.1
onnxruntime-gpu==1.20.1
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
proglog==0.1.10
propcache==0.2.0
protobuf==5.28.3
pydantic==2.9.2
pydantic_core==2.23.4
PyMatting==1.1.13
PyMySQL==1.1.1
python-dotenv==1.0.1
referencing==0.35.1
rembg==2.0.60
requests==2.32.3
rpds-py==0.21.0
scikit-image==0.24.0
scipy==1.14.1
setuptools==75.6.0
sympy==1.13.3
tifffile==2024.9.20
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.17.2
